Tidy up, use real message body
