+++ packages/update/src/sqs-event-types.ts
@@ -1,0 +1,17 @@
+import { z } from 'zod';
+import { stringToJSONSchema } from './zod-string-to-json';
+import { TranscriptionOutput } from '@guardian/transcription-service-common';
+
+export const SQSMessageBody = z.object({
+	MessageId: z.string(),
+	Timestamp: z.string(),
+	Message: stringToJSONSchema.pipe(TranscriptionOutput),
+});
+
+export const IncomingSQSEvent = z.object({
+	Records: z.array(
+		z.object({
+			body: stringToJSONSchema.pipe(SQSMessageBody),
+		}),
+	),
+});
